+++ rollup.config.js
@@ -1,6 +1,11 @@
-import json from '@rollup/plugin-json';
-const external = ['node:fs/promises', 'commander', 'node:path', '@shgysk8zer0/npm-utils/yaml.js', '@shgysk8zer0/npm-utils/json.js', '@shgysk8zer0/npm-utils/path.js', '@shgysk8zer0/polyfills'];
-const plugins = [json({ preferConst: true })];
+import { rollupImport } from '@shgysk8zer0/rollup-import';
+
+const external = [
+	'node:fs/promises', 'commander', 'node:path', '@shgysk8zer0/npm-utils/yaml.js',
+	'@shgysk8zer0/npm-utils/json.js', '@shgysk8zer0/npm-utils/path.js', '@shgysk8zer0/polyfills',
+];
+
+const plugins = [rollupImport()];
 
 export default [{
 	input: 'index.mjs',
@@ -28,9 +33,22 @@ export default [{
 	input: 'cli.mjs',
 	output: [{
 		file: 'cli.js',
+		banner: '#!/usr/bin/env node',
 		format: 'esm'
 	}, {
 		file: 'cli.cjs',
+		format: 'cjs'
+	}],
+	external,
+	plugins,
+}, {
+	input: 'html.mjs',
+	output: [{
+		file: 'html.js',
+		banner: '#!/usr/bin/env node',
+		format: 'esm'
+	}, {
+		file: 'html.cjs',
 		format: 'cjs'
 	}],
 	external,